basic impl

+++ plugin-audio/android/src/main/java/com/gluonhq/charm/down/plugins/android/AndroidAudioService.java
@@ -1,5 +1,9 @@
 package com.gluonhq.charm.down.plugins.android;
 
+import android.media.AudioAttributes;
+import android.media.AudioManager;
+import android.media.MediaPlayer;
+import android.media.SoundPool;
 import com.gluonhq.charm.down.Services;
 import com.gluonhq.impl.charm.down.plugins.Audio;
 import com.gluonhq.charm.down.plugins.AudioService;
@@ -10,15 +14,22 @@ import com.gluonhq.impl.charm.down.plugins.android.AndroidSound;
 
 import java.io.*;
 import java.net.URL;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * @author Almas Baimagambetov (almaslvl@gmail.com)
  */
 public class AndroidAudioService implements AudioService {
 
+    private SoundPool pool;
+
+    // TODO: use Charm Down cache?
+    private Map<String, Audio> cache = new HashMap<>();
     private File privateStorage;
 
     // TODO: web resource
+    // TODO: handle exceptions
     @Override
     public Audio createAudio(AudioType type, String resourceName) {
         if (privateStorage == null) {
@@ -27,26 +38,58 @@ public class AndroidAudioService implements AudioService {
 
         String subDirName = type == AudioType.MUSIC ? "music/" : "sounds/";
 
-//        System.out.println(url.toString());
-//        System.out.println(url.getFile());
-//        System.out.println(url.toExternalForm());
-
         String fileName = resourceName.substring(resourceName.lastIndexOf("/")+1, resourceName.length());
 
-        File outputFile = new File(privateStorage.getAbsolutePath() + "/assets/" + subDirName + fileName);
+        // assume this is unique, is it?
+        String fullName = privateStorage.getAbsolutePath() + "/assets/" + subDirName + fileName;
 
-        boolean result = copyFile(getClass().getResource(resourceName), outputFile);
+        if (cache.containsKey(fullName)) {
+            return cache.get(fullName);
+        }
 
-        System.out.println("Result: " + result);
+        File outputFile = new File(fullName);
 
-        if (type == AudioType.MUSIC) {
-            try {
-                return new AndroidMusic(new FileInputStream(outputFile), getClass().getResource(resourceName));
-            } catch (FileNotFoundException e) {
-                throw new RuntimeException(e);
+        if (!outputFile.exists()) {
+            copyFile(getClass().getResource(resourceName), outputFile);
+        }
+
+        try {
+            Audio audio;
+
+            FileInputStream stream = new FileInputStream(outputFile);
+
+            if (type == AudioType.MUSIC) {
+                MediaPlayer mediaPlayer = new MediaPlayer();
+                mediaPlayer.setAudioStreamType(AudioManager.STREAM_MUSIC);
+                mediaPlayer.setDataSource(stream.getFD());
+
+                mediaPlayer.prepare();
+
+                audio = new AndroidMusic(fullName, mediaPlayer);
+            } else {
+
+                if (pool == null) {
+                    AudioAttributes audioAttributes = new AudioAttributes.Builder()
+                            .setUsage(AudioAttributes.USAGE_GAME)
+                            .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
+                            .build();
+
+                    pool = new SoundPool.Builder()
+                            .setAudioAttributes(audioAttributes)
+                            .setMaxStreams(5)
+                            .build();
+                }
+
+                audio = new AndroidSound(fullName, pool, pool.load(stream.getFD(), 0, outputFile.length(), 1));
             }
-        } else {
-            return new AndroidSound(getClass().getResource(resourceName));
+
+            stream.close();
+
+            cache.put(fullName, audio);
+
+            return audio;
+        } catch (Exception e) {
+            throw new RuntimeException(e);
         }
     }
 
@@ -73,7 +116,8 @@ public class AndroidAudioService implements AudioService {
 
     @Override
     public void releaseAudio(Audio audio) {
-
+        cache.remove(audio.getFullName());
+        audio.dispose();
     }
 
     private boolean copyFile(URL fileURL, File outputFile)  {
@@ -97,84 +141,4 @@ public class AndroidAudioService implements AudioService {
         }
         return false;
     }
-//
-//    private SoundPool pool;
-//
-//    private Map<String, Integer> soundsCache = new HashMap<>();
-//
-//    @Override
-//    public void playSound(String fileName, int numTimes, double leftVolume, double rightVolume) {
-//        System.out.println("Play sound: " + fileName + " num times: " + numTimes);
-//
-//        if (pool == null) {
-//            System.out.println("Building pool");
-//            SoundPool.Builder builder = new SoundPool.Builder();
-//
-//            AudioAttributes audioAttributes = new AudioAttributes.Builder()
-//                    .setUsage(AudioAttributes.USAGE_GAME)
-//                    .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
-//                    .build();
-//
-//            builder.setAudioAttributes(audioAttributes)
-//                    .setMaxStreams(5);
-//
-//            pool = builder.build();
-//        }
-//
-//        if (soundsCache.containsKey(fileName)) {
-//            int id = soundsCache.get(fileName);
-//
-//            int streamID = pool.play(id, (float)leftVolume, (float)rightVolume, 1, numTimes - 1, 1.0f);
-//
-//            System.out.println("playing: " + streamID);
-//        } else {
-//
-//            try {
-//                if (privateStorage == null) {
-//                    privateStorage = Services.get(StorageService.class)
-//                            .flatMap(service -> service.getPrivateStorage())
-//                            .orElseThrow(() -> new RuntimeException("Error accesing Private Storage folder"));
-//                }
-//
-//                File assetsDir = new File(privateStorage, "assets");
-//                if (!assetsDir.exists()) {
-//                    System.out.println("CREATING MKDIR");
-//                    assetsDir.mkdir();
-//                }
-//
-//                File soundsDir = new File(assetsDir, "sounds");
-//                if (!soundsDir.exists()) {
-//                    System.out.println("CREATING SOUNDS DIR");
-//                    soundsDir.mkdir();
-//                }
-//
-//                File file = new File(privateStorage.getAbsolutePath() + fileName);
-//
-//                if (!file.exists()) {
-//                    boolean result = copyFile(fileName, file);
-//
-//                    System.out.println("COPYING: " + result);
-//                }
-//
-//                FileInputStream stream = new FileInputStream(file);
-//
-//                int soundID = pool.load(stream.getFD(), 0, file.length(), 1);
-//
-//                soundsCache.put(fileName, soundID);
-//
-//                stream.close();
-//
-//
-//                int streamID = pool.play(soundID, (float)leftVolume, (float)rightVolume, 1, numTimes - 1, 1.0f);
-//
-//                System.out.println("playing: " + streamID);
-//
-//
-//            } catch (Exception e) {
-//                e.printStackTrace(System.out);
-//            }
-//
-//
-//        }
-//    }
 }
